Fix following after plane teleports
